Removing Cyclic Redundancy and unnecesary imports
Cyclic Redundancy caused by MODULE_LIST, SCRIPT_LIST, ENGINE_LIST removed.
