+++ retriever/engines/csvengine.py
@@ -1,13 +1,10 @@
-from builtins import str
 from builtins import object
 
 import os
-import io
-import sys
-import csv
 
 from retriever.lib.models import Engine
-from retriever import DATA_DIR, open_fw, open_csvw
+from retriever.lib.scripts import open_fw, open_csvw
+from retriever.lib.defaults import DATA_DIR
 from retriever.lib.tools import sort_csv
 from retriever.lib.dummy import DummyConnection, DummyCursor
