+++ retriever/engines/jsonengine.py
@@ -1,16 +1,15 @@
 """Engine for writing data to a JSON file"""
-from builtins import zip
-from builtins import object
-from builtins import range
-
 import os
 import json
 
-from retriever.lib.models import Engine
-from retriever import DATA_DIR, open_fw, open_fr
+from builtins import zip
+from builtins import object
 from collections import OrderedDict
+from retriever.lib.models import Engine
+from retriever.lib.scripts import open_fw, open_fr
 from retriever.lib.tools import json2csv, sort_csv
 from retriever.lib.dummy import DummyConnection, DummyCursor
+from retriever.lib.defaults import DATA_DIR
 
 
 class engine(Engine):
