+++ retriever/engines/postgres.py
@@ -1,6 +1,6 @@
 import os
 from retriever.lib.models import Engine, no_cleanup
-from retriever import ENCODING
+from retriever.lib.defaults import ENCODING
 
 
 class engine(Engine):
