+++ retriever/engines/xmlengine.py
@@ -1,12 +1,11 @@
 import os
 
-from builtins import str
 from builtins import object
-from builtins import range
 from retriever.lib.models import Engine
-from retriever import DATA_DIR, open_fr, open_fw
 from retriever.lib.tools import xml2csv, sort_csv
 from retriever.lib.dummy import DummyConnection, DummyCursor
+from retriever.lib.scripts import open_fr, open_fw
+from retriever.lib.defaults import DATA_DIR
 
 
 class engine(Engine):
