+++ retriever/lib/scripts.py
@@ -1,0 +1,117 @@
+import os
+import csv
+import io
+from os.path import join, isfile, getmtime, exists
+from pkg_resources import parse_version
+import imp
+import sys
+
+from retriever.lib.defaults import SCRIPT_SEARCH_PATHS, VERSION, ENCODING
+from retriever.lib.compile import compile_json
+
+
+def MODULE_LIST(force_compile=False):
+    """Load scripts from scripts directory and return list of modules."""
+    modules = []
+    loaded_scripts = []
+
+    for search_path in [search_path for search_path in SCRIPT_SEARCH_PATHS if exists(search_path)]:
+        to_compile = [
+            file for file in os.listdir(search_path) if file[-5:] == ".json" and
+            file[0] != "_" and (
+                (not isfile(join(search_path, file[:-5] + '.py'))) or (
+                    isfile(join(search_path, file[:-5] + '.py')) and (
+                        getmtime(join(search_path, file[:-5] + '.py')) < getmtime(
+                            join(search_path, file)))) or force_compile)]
+        for script in to_compile:
+            script_name = '.'.join(script.split('.')[:-1])
+            compile_json(join(search_path, script_name))
+
+        files = [file for file in os.listdir(search_path)
+                 if file[-3:] == ".py" and file[0] != "_" and
+                 '#retriever' in ' '.join(open(join(search_path, file), 'r').readlines()[:2]).lower()]
+
+        for script in files:
+            script_name = '.'.join(script.split('.')[:-1])
+            if script_name not in loaded_scripts:
+                loaded_scripts.append(script_name)
+                file, pathname, desc = imp.find_module(script_name, [search_path])
+                try:
+                    new_module = imp.load_module(script_name, file, pathname, desc)
+                    if hasattr(new_module.SCRIPT, "retriever_minimum_version"):
+                        # a script with retriever_minimum_version should be loaded
+                        # only if its compliant with the version of the retriever
+                        if not parse_version(VERSION) >=  parse_version("{}".format(
+                                new_module.SCRIPT.retriever_minimum_version)):
+                            print("{} is supported by Retriever version {}".format(script_name,
+                                                                                   new_module.SCRIPT.retriever_minimum_version))
+                            print("Current version is {}".format(VERSION))
+                            continue
+                    # if the script wasn't found in an early search path
+                    # make sure it works and then add it
+                    new_module.SCRIPT.download
+                    modules.append(new_module)
+                except Exception as e:
+                    sys.stderr.write("Failed to load script: %s (%s)\nException: %s \n" % (
+                        script_name, search_path, str(e)))
+    return modules
+
+
+def SCRIPT_LIST(force_compile=False):
+    return [module.SCRIPT for module in MODULE_LIST(force_compile)]
+
+
+def open_fr(file_name, encoding=ENCODING, encode=True):
+    """Open file for reading respecting Python version and OS differences
+
+    Sets newline to Linux line endings on Windows and Python 3
+    When encode=False does not set encoding on nix and Python 3 to keep as bytes
+    """
+    if sys.version_info >= (3, 0, 0):
+        if os.name == 'nt':
+            file_obj = io.open(file_name, 'r', newline='', encoding=encoding)
+        else:
+            if encode:
+                file_obj = io.open(file_name, "r", encoding=encoding)
+            else:
+                file_obj = io.open(file_name, "r")
+    else:
+        file_obj = io.open(file_name, "r", encoding=encoding)
+    return file_obj
+
+
+def open_fw(file_name, encoding=ENCODING, encode=True):
+    """Open file for writing respecting Python version and OS differences
+
+    Sets newline to Linux line endings on Python 3
+    When encode=False does not set encoding on nix and Python 3 to keep as bytes
+    """
+    if sys.version_info >= (3, 0, 0):
+        if encode:
+            file_obj = io.open(file_name, 'w', newline='', encoding=encoding)
+        else:
+            file_obj = io.open(file_name, 'w', newline='')
+    else:
+        file_obj = io.open(file_name, 'wb',)
+    return file_obj
+
+
+def open_csvw(csv_file, encode=True):
+    """Open a csv writer forcing the use of Linux line endings on Windows
+
+    Also sets dialect to 'excel' and escape characters to '\\'
+
+    """
+    if os.name == 'nt':
+        csv_writer = csv.writer(csv_file, dialect='excel', escapechar='\\', lineterminator='\n')
+    else:
+        csv_writer = csv.writer(csv_file, dialect='excel', escapechar='\\')
+    return csv_writer
+
+
+def to_str(object, object_encoding=sys.stdout):
+    if sys.version_info >= (3, 0, 0):
+        enc = object_encoding.encoding
+        return str(object).encode(enc, errors='backslashreplace').decode("latin-1")
+    else:
+        return object
