+++ retriever/term_size.py
@@ -9,7 +9,6 @@ def get_terminal_size():
             import fcntl
             import termios
             import struct
-            import os
             cr = struct.unpack('hh', fcntl.ioctl(fd, termios.TIOCGWINSZ,
                                                  '1234'))
         except:
