+++ scripts/forest_plots_wghats.py
@@ -3,7 +3,7 @@
 
 from retriever.lib.models import Table, Cleanup, correct_invalid_value
 from retriever.lib.templates import Script
-from retriever import VERSION
+from retriever.lib.defaults import VERSION
 from pkg_resources import parse_version
