+++ scripts/vertnet_birds.py
@@ -6,7 +6,7 @@ from builtins import str
 from retriever.lib.models import Table
 from retriever.lib.templates import Script
 import os
-from retriever import VERSION
+from retriever.lib.defaults import VERSION
 from pkg_resources import parse_version
