+++ scripts/wood_density.py
@@ -11,8 +11,8 @@ from imp import reload
 
 from retriever.lib.templates import Script
 from retriever.lib.models import Table
-from retriever.lib.excel import Excel
-from retriever import HOME_DIR, open_fr, open_fw, open_csvw, to_str, VERSION
+from retriever.lib.defaults import VERSION
+from retriever.lib.scripts import open_fw, open_csvw, to_str
 from pkg_resources import parse_version
